+++ src/apps/PreferencesWindow/src/AppDelegate.m
@@ -1,8 +1,8 @@
 #import "AppDelegate.h"
 #import "ComplexModificationsFileImportWindowController.h"
+#import "DextAlertWindowController.h"
 #import "InputMonitoringAlertWindowController.h"
 #import "KarabinerKit/KarabinerKit.h"
-#import "KextdAlertWindowController.h"
 #import "PreferencesWindowController.h"
 #import "SimpleModificationsTableViewController.h"
 #import "SystemPreferencesManager.h"
@@ -12,7 +12,7 @@
 
 @property(weak) IBOutlet SimpleModificationsTableViewController* simpleModificationsTableViewController;
 @property(weak) IBOutlet ComplexModificationsFileImportWindowController* complexModificationsFileImportWindowController;
-@property(weak) IBOutlet KextdAlertWindowController* kextdAlertWindowController;
+@property(weak) IBOutlet DextAlertWindowController* dextAlertWindowController;
 @property(weak) IBOutlet InputMonitoringAlertWindowController* inputMonitoringAlertWindowController;
 @property(weak) IBOutlet NSWindow* preferencesWindow;
 @property(weak) IBOutlet PreferencesWindowController* preferencesWindowController;
@@ -39,7 +39,7 @@
 
   [self.preferencesWindowController setup];
 
-  [self.kextdAlertWindowController setup];
+  [self.dextAlertWindowController setup];
   [self.inputMonitoringAlertWindowController setup];
 }