+++ src/common.cmake
@@ -1,11 +1,13 @@
 # Common configurations
 
 set(CMAKE_OSX_DEPLOYMENT_TARGET "10.15" CACHE STRING "Minimum OS X deployment version")
+set(CMAKE_OSX_ARCHITECTURES "arm64;x86_64")
 
 set(CMAKE_CXX_STANDARD 17)
 
 include_directories(SYSTEM ${CMAKE_CURRENT_LIST_DIR}/vendor)
 include_directories(SYSTEM ${CMAKE_CURRENT_LIST_DIR}/vendor/cget/include)
+include_directories(SYSTEM ${CMAKE_CURRENT_LIST_DIR}/vendor/Karabiner-DriverKit-VirtualHIDDevice/include)
 include_directories(${CMAKE_CURRENT_LIST_DIR}/share)
 
 add_compile_options(-Wall)