+++ src/lib/libkrbn/src/libkrbn.cpp
@@ -12,6 +12,7 @@
 #include <fstream>
 #include <iostream>
 #include <nlohmann/json.hpp>
+#include <pqrs/karabiner/driverkit/virtual_hid_device_service.hpp>
 #include <string>
 
 namespace {
@@ -125,6 +126,10 @@ void libkrbn_launch_preferences(void) {
 
 void libkrbn_launch_multitouch_extension(void) {
   krbn::application_launcher::launch_multitouch_extension(false);
+}
+
+bool libkrbn_driver_running(void) {
+  return pqrs::karabiner::driverkit::virtual_hid_device_service::utility::driver_running();
 }
 
 bool libkrbn_system_core_configuration_file_path_exists(void) {
@@ -312,24 +317,6 @@ void libkrbn_enable_connected_devices_monitor(libkrbn_connected_devices_monitor_
 void libkrbn_disable_connected_devices_monitor(void) {
   if (libkrbn_components_manager_) {
     libkrbn_components_manager_->disable_connected_devices_monitor();
-  }
-}
-
-//
-// kextd_state_json_file_monitor
-//
-
-void libkrbn_enable_kextd_state_json_file_monitor(libkrbn_file_monitor_callback callback,
-                                                  void* refcon) {
-  if (libkrbn_components_manager_) {
-    libkrbn_components_manager_->enable_kextd_state_json_file_monitor(callback,
-                                                                      refcon);
-  }
-}
-
-void libkrbn_disable_kextd_state_json_file_monitor(void) {
-  if (libkrbn_components_manager_) {
-    libkrbn_components_manager_->disable_kextd_state_json_file_monitor();
   }
 }