+++ src/scripts/uninstall.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
-/bin/sh '/Library/Application Support/org.pqrs/Karabiner-Elements/uninstall_core.sh'
+/bin/bash '/Library/Application Support/org.pqrs/Karabiner-Elements/uninstall_core.sh'
+/bin/bash '/Library/Application Support/org.pqrs/Karabiner-DriverKit-VirtualHIDDevice/scripts/uninstall.sh'
 
 /usr/bin/killall karabiner_grabber
 /usr/bin/killall karabiner_console_user_server