+++ src/scripts/uninstall_core.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 PATH=/bin:/sbin:/usr/bin:/usr/sbin
 export PATH
@@ -30,20 +30,25 @@ chflags nouchg,noschg /Applications/Karabiner-EventViewer.app
 # Uninstall
 #
 
-bash '/Library/Application Support/org.pqrs/Karabiner-Elements/scripts/uninstall-Karabiner-VirtualHIDDevice.sh'
 rm -f '/Library/LaunchDaemons/org.pqrs.karabiner.karabiner_kextd.plist'
 rm -f '/Library/LaunchDaemons/org.pqrs.karabiner.karabiner_grabber.plist'
 rm -f '/Library/LaunchDaemons/org.pqrs.karabiner.karabiner_observer.plist'
+rm -f '/Library/LaunchAgents/org.pqrs.karabiner.agent.karabiner_grabber.plist'
+rm -f '/Library/LaunchAgents/org.pqrs.karabiner.agent.karabiner_observer.plist'
 rm -f '/Library/LaunchAgents/org.pqrs.karabiner.karabiner_console_user_server.plist'
 rm -f '/Library/LaunchAgents/org.pqrs.karabiner.karabiner_session_monitor.plist'
 rm -rf '/Applications/Karabiner-Elements.app'
 rm -rf '/Applications/Karabiner-EventViewer.app'
 rm -rf '/Library/Application Support/org.pqrs/Karabiner-Elements'
 
-#
-# Remove '/Library/StagedExtensions/Library/Application Support/org.pqrs/Karabiner-VirtualHIDDevice'
-#
+if [ -f '/Library/Application Support/org.pqrs/Karabiner-Elements/scripts/uninstall-Karabiner-VirtualHIDDevice.sh' ]; then
+    bash '/Library/Application Support/org.pqrs/Karabiner-Elements/scripts/uninstall-Karabiner-VirtualHIDDevice.sh'
 
-kextcache -prune-staging
+    #
+    # Remove '/Library/StagedExtensions/Library/Application Support/org.pqrs/Karabiner-VirtualHIDDevice'
+    #
+
+    kextcache -prune-staging
+fi
 
 exit 0