+++ tests/src/virtual_hid_device_utility/src/virtual_hid_device_utility_test.cpp
@@ -2,11 +2,13 @@
 
 #include "virtual_hid_device_utility.hpp"
 
-TEST_CASE("karabiner_virtual_hid_device::hid_report::modifiers") {
+TEST_CASE("pqrs::karabiner::driverkit::virtual_hid_device_driver::hid_report::modifiers") {
   {
-    pqrs::karabiner_virtual_hid_device::hid_report::modifiers modifiers;
-    modifiers.insert(pqrs::karabiner_virtual_hid_device::hid_report::modifier::left_shift);
-    modifiers.insert(pqrs::karabiner_virtual_hid_device::hid_report::modifier::right_command);
+    namespace hid_report = pqrs::karabiner::driverkit::virtual_hid_device_driver::hid_report;
+
+    hid_report::modifiers modifiers;
+    modifiers.insert(hid_report::modifier::left_shift);
+    modifiers.insert(hid_report::modifier::right_command);
 
     nlohmann::json expected = nlohmann::json::array();
     expected.push_back("left_shift");
@@ -17,8 +19,11 @@ TEST_CASE("karabiner_virtual_hid_device::hid_report::modifiers") {
 }
 
 TEST_CASE("karabiner_virtual_hid_device::hid_report::keys") {
+  namespace hid_report = pqrs::karabiner::driverkit::virtual_hid_device_driver::hid_report;
+
+  // keyboard_or_keypad
   {
-    pqrs::karabiner_virtual_hid_device::hid_report::keys keys;
+    hid_report::keys keys;
     keys.insert(type_safe::get(*(krbn::make_hid_usage(krbn::key_code::keyboard_a))));
     keys.insert(type_safe::get(*(krbn::make_hid_usage(krbn::key_code::keyboard_b))));
     keys.insert(type_safe::get(*(krbn::make_hid_usage(krbn::key_code::keyboard_c))));
@@ -30,8 +35,9 @@ TEST_CASE("karabiner_virtual_hid_device::hid_report::keys") {
 
     REQUIRE(krbn::virtual_hid_device_utility::to_json(keys, pqrs::hid::usage_page::keyboard_or_keypad).dump() == expected.dump());
   }
+  // consumer
   {
-    pqrs::karabiner_virtual_hid_device::hid_report::keys keys;
+    hid_report::keys keys;
     keys.insert(type_safe::get(*(krbn::make_hid_usage(krbn::consumer_key_code::rewind))));
     keys.insert(type_safe::get(*(krbn::make_hid_usage(krbn::consumer_key_code::eject))));
     keys.insert(type_safe::get(*(krbn::make_hid_usage(krbn::consumer_key_code::mute))));
@@ -43,8 +49,9 @@ TEST_CASE("karabiner_virtual_hid_device::hid_report::keys") {
 
     REQUIRE(krbn::virtual_hid_device_utility::to_json(keys, pqrs::hid::usage_page::consumer).dump() == expected.dump());
   }
+  // apple_vendor_top_case
   {
-    pqrs::karabiner_virtual_hid_device::hid_report::keys keys;
+    hid_report::keys keys;
     keys.insert(type_safe::get(pqrs::hid::usage::apple_vendor_top_case::keyboard_fn));
     keys.insert(type_safe::get(pqrs::hid::usage::apple_vendor_top_case::brightness_up));
     keys.insert(type_safe::get(pqrs::hid::usage::apple_vendor_top_case::illumination_up));
@@ -56,8 +63,9 @@ TEST_CASE("karabiner_virtual_hid_device::hid_report::keys") {
 
     REQUIRE(krbn::virtual_hid_device_utility::to_json(keys, pqrs::hid::usage_page::apple_vendor_top_case).dump() == expected.dump());
   }
+  // apple_vendor_keyboard
   {
-    pqrs::karabiner_virtual_hid_device::hid_report::keys keys;
+    hid_report::keys keys;
     keys.insert(type_safe::get(pqrs::hid::usage::apple_vendor_keyboard::expose_all));
     keys.insert(type_safe::get(pqrs::hid::usage::apple_vendor_keyboard::launchpad));
 
@@ -71,7 +79,7 @@ TEST_CASE("karabiner_virtual_hid_device::hid_report::keys") {
 
 TEST_CASE("karabiner_virtual_hid_device::hid_report::buttons") {
   {
-    pqrs::karabiner_virtual_hid_device::hid_report::buttons buttons;
+    pqrs::karabiner::driverkit::virtual_hid_device_driver::hid_report::buttons buttons;
     buttons.insert(1);
     buttons.insert(32);